feat(install_schema.sh): add script to install database schema from SQL files
The install_schema.sh script is added to the project. This script sets
the PGPASSWORD environment variable to the value of the PASS variable.
It then loops through all the SQL files in the /usr/src/appdata/
directory and executes them using the psql command, passing the
appropriate database connection parameters. This script is useful for
installing the database schema from SQL files.